Use TUnit's capabilities of Before(Test) to avoid having to use inheritance with a base call.
