+++ tests/Tests.InfiniLore.Database.Repositories/Scopes/LoreScopeQueryRepositoryTest.cs
